create map lookup table
- change delta time to double
- create lookup table for map
- render map to screen according to array of numbers

+++ rpg-game/Bullet.cpp
@@ -22,9 +22,9 @@ void Bullet::Load()
 {
 }
 
-void Bullet::Update(float deltaTime)
+void Bullet::Update(double deltaTime)
 {
-	rectangleShape.setPosition(rectangleShape.getPosition() + direction * m_speed * deltaTime);
+	rectangleShape.setPosition(rectangleShape.getPosition() + direction * m_speed * (float)deltaTime);
 }
 
 void Bullet::Draw(sf::RenderWindow& window)
