+++ rpg-game/Player.h
@@ -9,7 +9,7 @@ class Player
 private:
 	sf::Texture texture;
 	std::vector<Bullet> bullets;
-	float fireRateTimer;
+	double fireRateTimer;
 	int maxFireRate;
 	float playerSpeed;
 
@@ -25,6 +25,6 @@ public:
 	~Player();
 	void Initialize();
 	void Load();
-	void Update(Skeleton &skeleton, float deltaTime, sf::Vector2f &mousePosition);
+	void Update(Skeleton &skeleton, double deltaTime, sf::Vector2f &mousePosition);
 	void Draw(sf::RenderWindow& window);
 };
