+++ rpg-game/Skeleton.cpp
@@ -54,7 +54,7 @@ void Skeleton::Load()
     }
 }
 
-void Skeleton::Update(float deltaTime)
+void Skeleton::Update(double deltaTime)
 {
     if (health > 0) {
